added stable pimplefoam case
buoyantPimpleFoam was having problems with small mesh (zplus in ~40 range) and the solver blew up. I was wondering if it was possible to do small mash calculations with pimplefoam at all. This case shows with the right BCs, it is definitely possible to have small mesh size (zplus ~40) and resolved turbulence.